refactor: move the prompt and entities to parser
