--- api/core/workflow/utils/structured_output/entities.py
@@ -1,16 +1,0 @@
-from enum import StrEnum
-
-
-class ResponseFormat(StrEnum):
-    """Constants for model response formats"""
-
-    JSON_SCHEMA = "json_schema"  # model's structured output mode. some model like gemini, gpt-4o,  support this mode.
-    JSON = "JSON"  # model's json mode. some model like claude support this mode.
-    JSON_OBJECT = "json_object"  # json mode's another alias. some model like deepseek-chat, qwen use this alias.
-
-
-class SpecialModelType(StrEnum):
-    """Constants for identifying model types"""
-
-    GEMINI = "gemini"
-    OLLAMA = "ollama"
