--- api/core/workflow/utils/structured_output/prompt.py
@@ -1,17 +1,0 @@
-STRUCTURED_OUTPUT_PROMPT = """You’re a helpful AI assistant. You could answer questions and output in JSON format.
-constraints:
-    - You must output in JSON format.
-    - Do not output boolean value, use string type instead.
-    - Do not output integer or float value, use number type instead.
-eg:
-    Here is the JSON schema:
-    {"additionalProperties": false, "properties": {"age": {"type": "number"}, "name": {"type": "string"}}, "required": ["name", "age"], "type": "object"}
-
-    Here is the user's question:
-    My name is John Doe and I am 30 years old.
-
-    output:
-    {"name": "John Doe", "age": 30}
-Here is the JSON schema:
-{{schema}}
-"""  # noqa: E501
